Deploy heroku

+++ app/controllers/backoffice/config/site_controller.rb
@@ -1,33 +1,61 @@
 class Backoffice::Config::SiteController < ApplicationController
+   before_action :authenticate_user! 
    before_action :set_site, only: [:edit, :update, :destroy]
    before_action :carregaDropdowns, only: [:index, :new , :create, :edit, :update]
-   before_action :validaracesso, only: [:index, :new , :create, :edit, :update]
 
    layout "backoffice"
-   before_action :authenticate_user!
+   
 
   def index
-    @sites = Site.where("name LIKE ?", "%#{params[:q]}%").page(params[:page]).per(5)
-    @sitesCount = Site.where("name LIKE ?", "%#{params[:q]}%")
-    @names_users = Site.names_users_from_site(current_user)
-  end
-
-  def new
-    @site = Site.new
-  end
-
-  def create
-    @site = SiteService.create(params_site, current_user.id)
-
-    unless @site.errors.any?
-      redirect_to edit_backoffice_config_user_path(current_user.id), notice: "Site cadastrado com sucesso, agora termine de cadastrar as informações do seu perfil"
+    unless(Profile.read_permit(3, current_user.id, current_user.site) == 0)
+      @sites = Site.where("name LIKE ?", "%#{params[:q]}%").page(params[:page]).per(5)
+      @sitesCount = Site.where("name LIKE ?", "%#{params[:q]}%")
+      @names_users = Site.names_users_from_site(current_user)
     else
-      render :new
+      redirect_to "/422.html"
     end
   end
 
+  def new
+    if(current_user.site.blank?)
+      @site = Site.new
+    else
+      unless(Profile.create_permit(3, current_user.id, current_user.site) == 0) &&
+        @site = Site.new
+      else
+        redirect_to "/422.html"
+      end
+    end
+  end
+
+  def create
+
+    begin
+
+      @site = SiteService.create(params_site, current_user.id)
+
+      unless @site.errors.any?
+        redirect_to edit_backoffice_config_user_path(current_user.id), 
+        notice: "Site criado com sucesso, finalize seu perfil depois faça login novamente"
+      else
+        render :new
+      end
+      
+    rescue StandardError => e
+
+      render :new
+      flash.now[:notice] = "Erro ao cadastrar site: #{e}"
+      
+    end
+
+  end
+
   def edit
+    unless(Profile.edit_permit(3, current_user.id, current_user.site) == 0)
      set_site
+    else
+      redirect_to "/422.html"
+    end
   end
 
   
@@ -41,13 +69,17 @@ class Backoffice::Config::SiteController < ApplicationController
   end
 
   def destroy
+    unless(Profile.erase_permit(3, current_user.id, current_user.site) == 0)
+      site_name = @site.name
 
-    site_name = @site.name
+      if @site.destroy
+        redirect_to backoffice_config_site_index_path, notice: "Site #{site_name} excluido com sucesso"
+      else
+        render :index
+      end
 
-    if @site.destroy
-      redirect_to backoffice_config_site_index_path, notice: "Site #{site_name} excluido com sucesso"
     else
-      render :index
+      redirect_to "/422.html"
     end
   end
 
@@ -56,14 +88,6 @@ class Backoffice::Config::SiteController < ApplicationController
     def carregaDropdowns
       @admins = User.all 
       @profiles = Profile.profile_views(current_user, current_user.site)
-    end
-
-    def validaracesso
-      unless(current_user.site.blank?)
-        if(Profile.read_permit(3, current_user.id, current_user.site) == 0)
-          redirect_to "/422.html"
-        end
-      end
     end
 
     def set_site
